Read defaults on runtime
